+++ contextwatch/monitor/__init__.py
@@ -2,6 +2,7 @@
 
 from contextwatch.monitor.context_tracker import ContextSnapshot, ContextSummary, ContextTracker
 from contextwatch.monitor.latency_tracker import LatencySnapshot, LatencySummary, LatencyTracker
+from contextwatch.monitor.memory_tracker import MemorySnapshot, MemorySummary, MemoryTracker
 
 __all__ = [
     "ContextSnapshot",
@@ -10,4 +11,8 @@ __all__ = [
     "LatencySnapshot",
     "LatencySummary",
     "LatencyTracker",
+    "MemorySnapshot",
+    "MemorySummary",
+    "MemoryTracker",
 ]
+